Inline `ExprNumberLiteral` formatting logic
