+++ crates/ruff_python_formatter/src/expression/mod.rs
@@ -56,7 +56,6 @@ pub(crate) mod expr_tuple;
 pub(crate) mod expr_unary_op;
 pub(crate) mod expr_yield;
 pub(crate) mod expr_yield_from;
-pub(crate) mod number;
 mod operator;
 pub(crate) mod parentheses;
 pub(crate) mod string;
